+++ src/MySocailApp.Application/Queries/QuestionAggregate/QuestionImageResponseDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace MySocailApp.Application.Queries.QuestionAggregate
 {
-    public record QuestionImageResponseDto(int Id, int QuestionId, string BlobName, float Height, float Width);
+    public record QuestionImageResponseDto(int Id, int QuestionId, string BlobName, double Height, double Width);
 }
