+++ src/MySocailApp.Application/Queries/SolutionAggregate/SolutionImageResponseDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace MySocailApp.Application.Queries.SolutionAggregate
 {
-    public record SolutionImageResponseDto(int Id,int SolutionId,string BlobName,float Height,float Width);
+    public record SolutionImageResponseDto(int Id,int SolutionId,string BlobName, double Height, double Width);
 }
