+++ src/MySocailApp.Domain/MessageAggregate/Entities/MessageImage.cs
@@ -2,11 +2,11 @@
 
 namespace MySocailApp.Domain.MessageAggregate.Entities
 {
-    public class MessageImage(string blobName, float height, float width) : Entity
+    public class MessageImage(string blobName, double height, double width) : Entity
     {
         public int MessageId { get; private set; }
         public string BlobName { get; private set; } = blobName;
-        public float Height { get; private set; } = height;
-        public float Width { get; private set; } = width;
+        public double Height { get; private set; } = height;
+        public double Width { get; private set; } = width;
     }
 }
