+++ src/MySocailApp.Domain/QuestionAggregate/Entities/QuestionImage.cs
@@ -6,17 +6,17 @@
         public DateTime CreatedAt { get; private set; }
         public int QuestionId { get; private set; }
         public string BlobName { get; private set; }
-        public float Height { get; private set; }
-        public float Width { get; private set; }
+        public double Height { get; private set; }
+        public double Width { get; private set; }
 
-        private QuestionImage(string blobName, float height, float width)
+        private QuestionImage(string blobName, double height, double width)
         {
             BlobName = blobName;
             Height = height;
             Width = width;
         }
 
-        public static QuestionImage Create(string blobName, float height, float width)
+        public static QuestionImage Create(string blobName, double height, double width)
             => new(blobName, height, width) { CreatedAt = DateTime.UtcNow };
     }
 }
